User_custom_ldap/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.2
django-anymail==10.1
django-auth-ldap==4.4.0
django-bootstrap3==23.4
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-ldapdb==1.5.1
django-soap==0.1.9
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.0
grpcio==1.62.1
grpcio-tools==1.62.1
idna==3.6
ldap3==2.9.1
lxml==5.2.0
mysqlclient==2.2.0
oauthlib==3.2.2
paramiko==3.3.1
pika==1.3.2
Pillow==10.0.0
protobuf==4.25.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python-ldap==3.4.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
spyne==2.14.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
colorama==0.4.6
xmltodict==0.9.2
